requirements.txt
wxPython
requests
python-dotenv